% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin2r2c.R
\name{basin2r2c}
\alias{basin2r2c}
\title{Write MESH basin data to r2c file}
\usage{
basin2r2c(basin, header, r2cFile = "")
}
\arguments{
\item{basin}{Required. The values defining the basin parameters, as read in by 
\code{r2c2basin}. Can either be and array or a raster brick.}

\item{header}{Required. The header lines from the basin r2c file, as read in by 
\code{r2c2basin}}

\item{r2cFile}{Required. The file to be written.}
}
\value{
If successful, returns \code{TRUE}. If unsucessful, returns 
\code{FALSE}
}
\description{
Write MESH basin data to r2c file
}
\examples{
\dontrun{
# read in basin
basin <- r2c2basin("MESH_drainage_database.r2c", values_only = FALSE, 
as_rasters = FALSE)
# write as another file
basin2r2c(basin$basin, basin$header_lines, "new_basin.r2c")
}
}
\seealso{
\code{\link{r2c2basin}}
}
\author{
Kevin Shook
}
