% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinPeakSWE.R
\name{basinPeakSWE}
\alias{basinPeakSWE}
\title{Find annual maximum SWE}
\usage{
basinPeakSWE(waterBalance)
}
\arguments{
\item{waterBalance}{Requried. Data frame of water balance variables as returned by
\code{read_MESH_OutputTimeseries_csv}.}
}
\value{
Returns a data frame of \code{year}, \code{date_max_SWE}, and \code{max_SWE}. 
Note that the SWE is the sum of the variables \code{SNO} and \code{WSNO}.
}
\description{
Find annual maximum SWE
}
\examples{
\dontrun{waterBalance <- read_MESH_OutputTimeseries_csv("Basin_average_water_balance.csv", 
missingValueThreshold = -1e6)
basinPeakSWE(waterBalance)}

}
