% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinPrecipEvapRunoffPlot.R
\name{basinPrecipEvapRunoffPlot}
\alias{basinPrecipEvapRunoffPlot}
\title{Plots basin precipitation evaporation and runoff}
\usage{
basinPrecipEvapRunoffPlot(basinWaterBalance)
}
\arguments{
\item{basinWaterBalance}{Required. Data frame to be plotted. As read in by \code{readOutputTimeseriesCSV}.}
}
\value{
Returns a \pkg{ggplot2} line plot of the variable values (mm).
}
\description{
Plots basin precipitation evaporation and runoff
}
\examples{
\dontrun{
waterBalance <- readOutputTimeseriesCSV("Basin_average_water_balance.csv")
p <- basinPrecipEvapRunoffPlot(waterBalance)}
}
\seealso{
\code{\link{readOutputTimeseriesCSV}} \code{\link{basinStoragePlot}} \code{\link{basinSoilWaterIcePlot}}
}
\author{
Kevin Shook
}
