% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinRunoffPlot.R
\name{basinRunoffPlot}
\alias{basinRunoffPlot}
\title{Plots basin runoff components}
\usage{
basinRunoffPlot(basinWaterBalance, cumul = FALSE)
}
\arguments{
\item{basinWaterBalance}{Required. Data frame to be plotted. As read in by 
\code{readOutputTimeseriesCSV}.}

\item{cumul}{Optional. If \code{FALSE}, then interval values are plotted. 
If \code{TRUE} (the default) then cumulative values are plotted. Note that the cumulative 
values are determined by summing the interval values \emph{not} by plotting the 
MESH cumulative variables.}
}
\value{
Returns a \pkg{ggplot2} line plot of the variable values (mm).
}
\description{
Plots basin runoff components
}
\examples{
\dontrun{
waterBalance <- readOutputTimeseriesCSV("Basin_average_water_balance.csv")
p <- basinRunoffPlot(waterBalance)}
}
\seealso{
\code{\link{readOutputTimeseriesCSV}} \code{\link{basinStoragePlot}} \code{\link{basinSoilWaterIcePlot}}
}
\author{
Kevin Shook
}
