% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinSoilWaterPlot.R
\name{basinSoilWaterPlot}
\alias{basinSoilWaterPlot}
\title{Plots basin soil liquid water}
\usage{
basinSoilWaterPlot(basinWaterBalance, layers = c(1, 2, 3, 4, 5, 6))
}
\arguments{
\item{basinWaterBalance}{Required. Data frame to be plotted. As read in by \code{readOutputTimeseriesCSV}.}

\item{layers}{Optional. A vector of the layers to be plotted. By default layers 1 though 6 are used.}
}
\value{
Returns a \pkg{ggplot2} line plot of the variable values (mm).
}
\description{
Plots basin soil liquid water
}
\examples{
\dontrun{
waterBalance <- readOutputTimeseriesCSV("Basin_average_water_balance.csv")
p <- basinSoilWaterPlot(waterBalance)}
}
\seealso{
\code{\link{readOutputTimeseriesCSV}} \code{\link{basinStoragePlot}} \code{\link{basinSoilWaterIcePlot}}
}
\author{
Kevin Shook
}
