% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinStoragePlot.R
\name{basinStoragePlot}
\alias{basinStoragePlot}
\title{Plots total basin storage}
\usage{
basinStoragePlot(basinWaterBalance)
}
\arguments{
\item{basinWaterBalance}{Required. Data frame to be plotted. As read in by 
\code{readOutputTimeseriesCSV}.}
}
\value{
Returns a \pkg{ggplot2} line plot of the value of \code{STG} (mm).
}
\description{
Plots total basin storage
}
\examples{
\dontrun{
waterBalance <- readOutputTimeseriesCSV("Basin_average_water_balance.csv")
p <- basinStoragePlot(waterBalance)}
}
\seealso{
\code{\link{readOutputTimeseriesCSV}} \code{\link{basinStorageVariablesPlot}} \code{\link{basinSoilWaterIcePlot}}
}
\author{
Kevin Shook
}
