% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinStorageVariablesPlot.R
\name{basinStorageVariablesPlot}
\alias{basinStorageVariablesPlot}
\title{Plots basin water balance storage components}
\usage{
basinStorageVariablesPlot(basinWaterBalance, varNames = "",
  layers = c(1, 2, 3, 4, 5, 6))
}
\arguments{
\item{basinWaterBalance}{Required. Data frame to be plotted. As read in by \code{read_MESH_OutputTimeseries_csv}.}

\item{varNames}{Optional. A vector of the names of the variables to be plotted. 
If not specified, all of the variables listed above will be plotted.}

\item{layers}{Optional. A vector of the layers to be plotted. By default 
layers 1 through 6 are used.}
}
\value{
Returns a \pkg{ggplot2} stacked area time plot of the variable values (mm).
}
\description{
Creates a \pkg{ggplot2} stacked area plot of specified water 
balance storage components. The variables plotted may include:
\describe{
  \item{variable}{definition}
  \item{SNCAN}{Snow component of precipitation intercepted by the canopy}
  \item{RCAN}{Rain component of precipitation intercepted by the canopy	mm or kg m-2 of water	AVG}
  \item{SNO}{Snow water equivalent (SWE) of the snow mass}
  \item{ZPND}{Depth of water ponded at the surface}
  \item{LQWS}{Water equivalent of the volumetric liquid water content of the soil}
  \item{FZWS}{Water equivalent of the volumetric frozen water content of the soil}
  \item{ALWS}{Water equivalent of the volumetric liquid and frozen water contents of the soil, sum of LQWS and FZWS}
  }
}
\examples{
\dontrun{
waterBalance <- read_MESH_OutputTimeseries_csv("Basin_average_water_balance.csv")
p <- basinStorageVariablesPlot(waterBalance)
# the plot can have a restricted date range
startDate <- as.Date("2005-10-01", format = "\%Y-\%m-\%d")
endDate <- as.Date("2006-09-30", format = "\%Y-\%m-\%d")
library(ggplot2)
p <- p + xlim(startDate, endDate)
p
# you can also change the colours used, either by
using a defined scale, or by manually specifying them
# This example uses the colours in the package viridis
# which scales from dark to light by reversing the direction
library(viridis)
p <- p +  scale_fill_viridis(discrete = TRUE, direction = -1)
p
}
}
\seealso{
\code{\link{read_MESH_OutputTimeseries_csv}}
}
\author{
Kevin Shook
}
