% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinWaterBalancePlot.R
\name{basinWaterBalancePlot}
\alias{basinWaterBalancePlot}
\title{Plots basin precipitation cumulative water balance.}
\usage{
basinWaterBalancePlot(basinWaterBalance, accumulate_delta_storage = TRUE)
}
\arguments{
\item{basinWaterBalance}{Required. Data frame to be plotted. As read in by 
\code{read_MESH_OutputTimeseries_csv}. Note that because the value of \code{DTSG} 
(delta storage) can be negative, you need to set a threshold value much smaller than
zero when you read in the values.}

\item{accumulate_delta_storage}{Optional. If \code{TRUE} (the default),
the delta storage is accumulated from the beginning of the data set. If
\code{FALSE}, the delta storage values in the file are \emph{not} accumulated
as they are assumed to be cumualtive values.}
}
\value{
Returns a \pkg{ggplot2} line plot of the variable values (mm).
}
\description{
As with the \code{basinRunoffPlot}, the cumulative values of
precipitation, evaporation and runoff (and optionally, delta storage) 
are computed by the function, rather than by using the MESH variables. This
allows the plot to be used on a sub-set of the basin output data.
}
\examples{
\dontrun{
waterBalance <- read_MESH_OutputTimeseries_csv("Basin_average_water_balance.csv", 
missingValueThreshold = -1e6)
p <- basinWaterBalancePlot(waterBalance)}
}
\seealso{
\code{\link{read_MESH_OutputTimeseries_csv}} \code{\link{basinStoragePlot}} \code{\link{basinSoilWaterIcePlot}}
}
\author{
Kevin Shook
}
