\name{category}
\docType{methods}

\alias{generic,category}
\alias{category}
\alias{category<-}
\alias{category,MeSHHyperGParams-method}
\alias{category<-,MeSHHyperGParams,character-method}

\title{A function to return the name of MeSH category}
\description{This function returns the name of MeSH category.}

\usage{
category(r)
category(r) <- value
}

\arguments{
  \item{r}{
    An object containing annotation information.
  }
  \item{value}{
    The annotation information to set on \code{object}.
  }
}
\author{Koki Tsuyuzaki}

\keyword{methods}
