\name{database}
\docType{methods}

\alias{generic,database}
\alias{database}
\alias{database<-}
\alias{database,MeSHHyperGParams-method}
\alias{database<-,MeSHHyperGParams,character-method}

\title{A function to return the name of MeSH database}
\description{This function returns the name of MeSH database.}

\usage{
database(r)
database(r) <- value
}

\arguments{
  \item{r}{
    An object containing annotation information.
  }
  \item{value}{
    The annotation information to set on \code{object}.
  }
}
\author{Koki Tsuyuzaki}

\keyword{methods}
