% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleHydrograph.R
\name{doubleHydrograph}
\alias{doubleHydrograph}
\title{Creates hydrograph from 2 MESH output files}
\usage{
doubleHydrograph(
  MESHvals1,
  stationNames1 = "",
  MESHname1 = "MESH1",
  MESHvals2,
  stationNames2 = "",
  MESHname2 = "MESH2",
  byStation = TRUE,
  byYear = FALSE,
  meas = TRUE,
  sim = TRUE,
  calStart = "",
  calEnd = "",
  alpha = 1
)
}
\arguments{
\item{MESHvals1}{Required. A data frame of output from a MESH run, as produced by \code{read_MESH_OutputTimeseries_csv}.}

\item{stationNames1}{Optional. A vector of strings holding station names. If specified, the station names will 
be used in the plots. Otherwise the MESH station numbers will be used.}

\item{MESHname1}{Optional. A string giving the name of the first MESH output. Default is "MESH1".}

\item{MESHvals2}{Required. A data frame of output from a MESH run, as produced by \code{read_MESH_OutputTimeseries_csv}.}

\item{stationNames2}{Optional. A vector of strings holding station names.}

\item{MESHname2}{Optional. A string giving the name of the second MESH output. Default is "MESH2".}

\item{byStation}{Optional. If \code{TRUE} (the default) then the plots will be coloured according to the station names. You may want to set this to \code{FALSE} if you are facetting by station name.}

\item{byYear}{Optional. If \code{TRUE} then the plots will be able to be facetted by year. Note that 
this means that the dates are all plotted using the year 2000, so you will see strange results
 if you set this to \code{TRUE} and don't facet by year. Default is \code{FALSE}}

\item{meas}{Optional. Should the measured values be plotted? Default is \code{TRUE}. If \code{FALSE}, they will be omitted.}

\item{sim}{Optional. Should the simulated values be plotted? Default is \code{TRUE}. If \code{FALSE}, they will be omitted.}

\item{calStart}{Optional. The start date of the calibration period. Must be a string in the format \option{yyyy-mm-dd}. If specified, values on and after this date will be designated as the \code{Calibration} period. The remaining values will be designated as the \code{Validation} period.}

\item{calEnd}{Optional. The start date of the calibration period. Must be a string in the format \option{yyyy-mm-dd}. If specified, values on and after this date will be designated as the \code{Calibration} period. The remaining values will be designated as the \code{Validation} period.}

\item{alpha}{Optional. Sets the alpha channel (transparency) of the plots. The default value is \code{1}, i.e. opaque. Setting \code{alpha} to less than 1 makes the plots transparent, which can be useful to see overlapping hydrographs.}
}
\value{
If successful, returns a \pkg{ggplot2} object. If unsuccessful, returns \code{FALSE}. The object can be facetted by the name of the station (the variable is called \code{station}). If the option \code{byYear = TRUE}, then the object can be facetted by the variable \code{YEAR}.
}
\description{
Creates a \pkg{ggplot} hydrograph from MESH output. This function \emph{only} uses 
values from two MESH data frames (as read in using \code{readOutputTimeseriesCSV}). It is assumed
that the observed data are the same in both cases - only the simulations differ.
Because this function returns a \pkg{ggplot} object, you can change its format in any way you like.
The plots produced may be faceted using the commands \code{facet_wrap} or \code{facet_grid}.
}
\note{
Specifying the calibration start and/or end dates will allow the resulting plot to be facetted by the variable \code{period}.
}
\examples{
\dontrun{
p <- doubleHydrograph(MESHvals1 = capa, MESHname1 = "CaPa precip", 
MESHvals2 = kevin, MESHname2 = "Station precip",  byStation = TRUE, 
alpha = 0.6)
p
# facet by station
p2 <- p + facet_wrap(~station, scales = "free_y", strip.position = "right") 
p2
# add scale colour
colours <- c("red", "blue", "black")
p3 <- p2 + + scale_colour_manual(values = colours)
# restrict plot to a range of dates
p3
startDate <- as.Date("2005-04-01")
endDate <- as.Date("2005-10-01")
p4 <- p3 + scale_x_date(limits = c(startDate, endDate))
p4
}

}
\seealso{
\code{\link{simpleHydrograph}} \code{\link{read_MESH_OutputTimeseries_csv}} \code{\link{hydroStats}}
}
\author{
Kevin Shook
}
