% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroStats.R
\name{hydroStats}
\alias{hydroStats}
\title{Calculates Goodness of Fit statistics for MESH output}
\usage{
hydroStats(MESHvals, stationNames = "")
}
\arguments{
\item{MESHvals}{Required. A data frame of output from a MESH run, as produced by \code{readOutputTimeseriesCSV}.}

\item{stationNames}{Optional. A vector of strings holding station names. If specified, the station names will 
be used in the returned data frame, otherwise the MESH station numbers will be used.}
}
\value{
Returns a data frame with the following variables:
#' \describe{
  \item{station}{station name or number}
  \item{me}{Mean Error}
  \item{mae}{Mean Absolute Error}
  \item{mse}{Mean Squared Error}
  \item{rmse}{Root Mean Square Error}
  \item{nrmse}{Normalized Root Mean Square Error  ( -100\% <= nrms <= 100\% )}
  \item{PBIAS}{Percent Bias }
  \item{pbiasfdc}{PBIAS in the slope of the midsegment of the Flow Duration Curve}
  \item{RSR}{Ratio of RMSE to the Standard Deviation of the Observations, RSR = rms / sd(obs). ( 0 <= RSR <= +Inf )}
  \item{rSD}{Ratio of Standard Deviations, rSD = sd(sim) / sd(obs)}
  \item{NSE}{Nash-Sutcliffe Efficiency ( -Inf <= NSE <= 1 )}
  \item{mNSE}{Modified Nash-Sutcliffe Efficiency}
  \item{rNSE}{Relative Nash-Sutcliffe Efficiency }
  \item{d}{Index of Agreement ( 0 <= d <= 1 )}
  \item{d1}{Modified Index of Agreement}
  \item{rd}{Relative Index of Agreement}
  \item{cp}{Persistence Index ( 0 <= PI <= 1 ) }
  \item{r}{Pearson Correlation coefficient ( -1 <= r <= 1 )}
  \item{r.Spearman}{Spearman Correlation coefficient ( -1 <= r.Spearman <= 1 ) }
  \item{R2}{Coefficient of Determination ( 0 <= R2 <= 1 ). \cr
    Gives the proportion of the variance of one variable that is predictable from the other variable}
  \item{bR2}{R2 multiplied by the coefficient of the regression line between \code{sim} and \code{obs} \cr ( 0 <= bR2 #'   <= 1 )}
  \item{KGE}{Kling-Gupta efficiency between \code{sim} and \code{obs} \cr ( 0 <= KGE <= 1 )}
  \item{VE}{Volumetric efficiency between \code{sim} and \code{obs} \cr ( -Inf <= VE <= 1)}
}
}
\description{
This function is a wrapper for the function \code{gof} in the package \pkg{hydroGOF}. 
It computes several Goodness of Fit statistics for each station. The output of this function may be
used on its own, or to annotate a hydrograph.
}
