% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseNums}
\alias{parseNums}
\title{Parses a string containing numbers}
\usage{
parseNums(numString)
}
\arguments{
\item{numString}{Required. A character string containing numbers separated by any number of spaces.}
}
\value{
Returns a numeric vector.
}
\description{
Parses a string containing numbers
}
\examples{
parseNums(' 1  2 3   4     5 ')
}
\author{
Kevin Shook
}
