% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseText}
\alias{parseText}
\title{Parses a string containing several sub-strings}
\usage{
parseText(textString)
}
\arguments{
\item{textString}{Required. A character string containing strings separated by any number of spaces.}
}
\value{
Returns a character vector.
}
\description{
Parses a string containing several sub-strings
}
\examples{
parseText(' red  green    blue      black')
}
\author{
Kevin Shook
}
