% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2c2raster.R
\name{r2c2raster}
\alias{r2c2raster}
\title{Reads r2c file to raster brick}
\usage{
r2c2raster(r2cFile, NAvalue = NULL, as_rts = FALSE, timezone = "",
  layerNameFormat = NULL)
}
\arguments{
\item{r2cFile}{Required. Name of \code{r2c} file containing time series.}

\item{NAvalue}{Optional. If specified, values smaller than \code{NAvalue}
will be set to \code{NA_real_}}

\item{as_rts}{Optional. If \code{TRUE}, the returned value will
be a \pkg{rts} object, which allows the creation of 1-D time series, and for
temporal aggregation. If \code{FALSE} (the default) a standard \pkg{raster} 
\code{brick} object is returned, which is better for simple plotting of
the layers.}

\item{timezone}{Optional. If the \code{r2cFile} contains date values for 
each Frame, then the Frame times are returned as \R dates. If there are 
hours and seconds, then they will be converted to \code{POSIXct} datetime
values. In this case, you may want to specify the \code{timezone} of the data.
If the \code{timezone} is not specified, your default value will be used.}

\item{layerNameFormat}{Optional. Sets the layer names when returning the
\pkg{raster} \code{brick} to avoid conflicting with the \R variable rules.}
}
\value{
Returns eiher a \pkg{raster} \code{brick} or an \pkg{rts} 
\code{rts} object.
}
\description{
This function reads a file
containing a time series of 2D values, which is output from
a MESH model. It is not intended to read in a file
describing a drainage basin. For that purpose, you should be using the 
function \code{r2c2basin}. This function returns eiher a \pkg{raster} \code{brick} or 
an \pkg{rts} 
\code{rts} object, which is a timeseries raster. Each Frame in the original 
file becomes a separate layer. 
'The name of each layer in the \pkg{raster} \code{brick} is set to the 
time stamp of each Frame. Because the layer names
are standard \R variables, they must obey the rules for variable names,
inclusing beginning with a character, and not containing spaces. These rule
will change the layer names if you are not careful.
}
\examples{
\dontrun{
temps <- r2c2raster("TA_M.r2c", NAvalue = 0, as_rts = FALSE, layerNameFormat = "\%b_\%Y")
# convert air temps from K to C
temps <- temps - 273.15
plot(temps)
# create an animation and save it as a file
library(animation)
saveGIF(animate(temps, n = 1))}
}
\seealso{
\code{\link{rts}} \code{\link{r2c2basin}}
}
\author{
Kevin Shook
}
