% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MESH_OutputTimeseries_csv.R
\name{read_MESH_OutputTimeseries_csv}
\alias{read_MESH_OutputTimeseries_csv}
\title{Reads MESH output .csv file containing timeseries}
\usage{
read_MESH_OutputTimeseries_csv(
  outputFile,
  timezone = "",
  missingValueThreshold = -0.1
)
}
\arguments{
\item{outputFile}{Required. Name of MESH output file. Must be a .csv file.}

\item{timezone}{Not required for daily time series. Required for sub-daily time series. The name of the timezone of the data as a character string. This should be the timezone of your data, but omitting daylight savings time. Note that the timezone code is specific to your OS. To avoid problems, you should use a timezone without daylight savings time. You can use \option{etc/GMT+6} or \option{etc/GMT+7} for Central Standard and Mountain Standard time, respectively. DO NOT use \option{America/Regina} as the time zone, as it includes historical changes between standard and daylight savings time.}

\item{missingValueThreshold}{Optional. Any value smaller than this value will be set to \code{NA_real_} when the file is imported. The default value is -0.1 to prevent zero values from being affected.}
}
\value{
If successful, returns a data frame. The first columns will be called 
\option{DATE} for daily values, and will contain a standard \R date. For sub-daily timeseries
the first column will be called \option{DATETIME} and will contain a standard \code{POSIXct} 
date/time. If unsuccessful, returns the value \code{FALSE}.
}
\description{
Reads a file containing any output from a MASH model into a standard \R data frame. 
The names of the varables will be trimmed to remove leading and trailing spaces, and the time
variables are combined into a single \R date or datetime.
}
\examples{
\dontrun{
timezone <- 'etc/GMT+6'
outfile <- "Basin_average_water_balance_ts.csv"
output <- read_MESH_OutputTimeseries_csv(outfile, timezone)}
}
\seealso{
\code{\link{simpleHydrograph}}
}
\author{
Kevin Shook
}
