% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tb02.R
\name{read_tb02}
\alias{read_tb02}
\title{Reads in a tb0 file}
\usage{
read_tb02(tb0File = "", values_only = TRUE, timezone = "",
  NAvalue = NULL)
}
\arguments{
\item{tb0File}{Required. The name of the file to be read.}

\item{values_only}{Optional. If \code{TRUE}, only the time series values will
be returned. If {FALSE}, the meta data will also be returned. Note that if
the value is set to \code{TRUE}, and there are no time series values in the
file, as for a reservoir file, then an error will result.}

\item{timezone}{Optional. The data time series have \code{POSIXct} datetime
values. You may want to specify the \code{timezone} of the data.
If the \code{timezone} is not specified, your default value (i.e. your
time zone) will be used.}

\item{NAvalue}{Optional. If specified, values smaller than \code{NAvalue}
will be set to \code{NA_real_}}
}
\value{
Returns time series as a data frame. If meta data are specified, they
are returned as lists.
}
\description{
Reads in a tb0 file. The time series values, if present, will be
stored in a data frame. Optionally, the meta data will be stored as lists. 
The meta data are of 2 types, the header values, which refer to the entire 
file and column values, which pertain to individual columns.
}
\examples{
\dontrun{
qvals <- read_tb02("MESH_input_streamflow.tb0", NAvalue = -0.01, values_only = TRUE)
}
}
\seealso{
\code{\link{read_r2c_raster}} \code{\link{read_MESH_OutputTimeseries_csv}}
}
\author{
Kevin Shook
}
