% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleHydrograph.R
\name{simpleHydrograph}
\alias{simpleHydrograph}
\title{Creates hydrograph from MESH output}
\usage{
simpleHydrograph(MESHvals, stationNames = "", byStation = TRUE,
  byYear = FALSE, meas = TRUE, sim = TRUE, stats = "")
}
\arguments{
\item{MESHvals}{Required. A data frame of output from a MESH run, as produced by \code{readOutputTimeseriesCSV}.}

\item{stationNames}{Optional. A vector of strings holding station names. If specified, the station names will 
be used in the plots. Otherwise the MESH station numbers will be used.}

\item{byStation}{Optional. If \code{TRUE} (the default) then the plots will be coloured according to the station names. You may want to set this to \code{FALSE} if you are facetting by station name.}

\item{byYear}{Optional. If \code{TRUE} then the plots will be able to be facetted by year. Note that this means that 
the dates are all plotted using the year 2000, so you will see strange results if you set this to \code{TRUE} and 
don't facet by year. Default is \code{FALSE}}

\item{meas}{Optional. Should the measured values be plotted? Default is \code{TRUE}. If \code{FALSE}, they will be omitted.}

\item{sim}{Optional. Should the simulated values be plotted? Default is \code{TRUE}. If \code{FALSE}, they will be omitted.}
}
\value{
If successful, returns a \pkg{ggplot2} object. If unsuccessful, returns \code{FALSE}. The object can be facetted by the name of the station (the variable is called \code{station}). If the option \code{byYear = TRUE}, then the object can be facetted by the variable \code{YEAR}.
}
\description{
Creates a \pkg{ggplot} hydrograph from MESH output. This function \emph{only} uses 
values from a single MESH data frame (as read in using \code{readOutputTimeseriesCSV}), so does
not work with outside sources, such as WSC files. Because this function returns a 
\pkg{ggplot} object, you can change its format in any way you like.
The plots produced may be faceted using the commands \code{facet_wrap} or \code{facet_grid}.
}
\examples{
# plot hydrograph of all data on single graph
p1 <- simpleHydrograph(MESH_streamflows)
p1
# add station names, and replot
stations <- c("Station1", "Station2")
p2 <- simpleHydrograph(MESH_streamflows, stationNames = stations)
p2
# remove colouring by station, and facet, changing the axis label format
p3 <- simpleHydrograph(MESH_streamflows, stationNames = stations, byStation = FALSE)
# load in all of ggplot2 to modify plots
library(ggplot2)
p3 <- p3 + facet_wrap(~station, nrow = 2) + scale_x_date(date_labels = "\%Y")
p3
# plot by year, then facet
p4 <- simpleHydrograph(MESH_streamflows, stationNames = stations, byYear = TRUE)
p4 <- p4 + facet_wrap(~YEAR, scales = "free_y")
p4
# remove colouring for stations, and facet by station and year
p5 <- simpleHydrograph(MESH_streamflows, stationNames = stations, byStation = FALSE, byYear = TRUE)
p5 <- p5 + facet_grid(YEAR~station, scales = "free_y")
# change colours
plotcols <- c("red", "blue")
p5 <- p5 + scale_colour_manual(values = plotcols)
p5
}
\author{
Kevin Shook
}
