% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_present}
\alias{var_present}
\title{Find if variable is present in a data frame}
\usage{
var_present(dataframe, variable)
}
\arguments{
\item{dataframe}{Required. Dataframe to be checked}

\item{variable}{Required. Variable name as a character string.}
}
\value{
Returns \code{TRUE} or \code{FALSE}
}
\description{
Find if variable is present in a data frame
}
\examples{
\dontrun{var_present(values, "datetime")}
}
