% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tb0.R
\name{write_tb0}
\alias{write_tb0}
\title{Writes a MESH tb0 file}
\usage{
write_tb0(values = NULL, column_meta = NULL, header = NULL,
  NAvalue = -1, tb0File = "")
}
\arguments{
\item{values}{Optional. A data frame of the values to be written, if the file
is to be a time seires. Note that the first column must be called 
\code{datetime} and must be a POSIXct date/time.}

\item{column_meta}{Required. A data frame containing the following columns
#' \describe{
 \item{columnUnits}{required}
 \item{columnType}{required}
 \item{columnName}{required}
 \item{columnLocationX}{required}
 \item{columnLocationY}{required}
 \item{DA}{optional}
 \item{coeff1}{optional}
 \item{coeff2}{optional}
 \item{coeff3}{optional}
 \item{coeff4}{optional}
 \item{coeff5}{optional}
}}

\item{header}{Required. A list containing the following variables
 \describe{
 \item{filetype}{optional, default is \code{tb0  ASCII  EnSim 1.0}}
 \item{datatype}{optional, default is \code{Time Series}}
 \item{application}{optional, default is \code{EnSimHydrologic}}
 \item{version}{optional, default is default is \code{2.1.23}}
 \item{written_by}{optional, default is default is \code{MESHr}}
 \item{creation_date}{optional, default is current date/time}
 \item{source_file}{optional, default is nothing}
 \item{name}{required}
 \item{projection}{required, character string}
 \item{ellipsoid}{required, character string}
 \item{start_time}{required if \code{values} are not specified}
 \item{delta_t}{required if \code{values} are not specified}
 \item{attributeunits}{optional, default is nothing}
 \item{unitconversion}{optional, default is nothing}
}}

\item{NAvalue}{Optional. Value to be used for \code{NA_real_} values in
the \code{.tb0} file. The default value is -1, which is not suitable
for air temperatures.}

\item{tb0File}{Required. Name of file to be written.}
}
\value{
If successful, returns \code{TRUE}. If unsuccessful, returns 
\code{FALSE}
}
\description{
Writes a MESH tb0 file
}
\note{
If you have read in the \code{header} data from another file, and you
want to use the default values, you can just replace the existing values with
a \code{NULL} value. You might want to do this to use the new start time, 
and creation date.
}
\examples{
\dontrun{ 
header$start_time <- NULL
header$creation_date <- NULL
write_tb0(values, column_meta_data, header, "MESH_values.ts0")
}
}
\seealso{
\code{\link{read_tb0}}
}
\author{
Kevin Shook
}
