% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinSnowPlot.R
\name{basinSnowPlot}
\alias{basinSnowPlot}
\title{Plots basin snow water equivalent}
\usage{
basinSnowPlot(basinWaterBalance)
}
\arguments{
\item{basinWaterBalance}{Required. Data frame to be plotted. As read in by \code{read_MESH_OutputTimeseries_csv}.}
}
\value{
Returns a \pkg{ggplot2} stacked line plot of the basin SWE (mm).
}
\description{
Plots basin snow water equivalent
}
\examples{
\dontrun{
waterBalance <- read_MESH_OutputTimeseries_csv("Basin_average_water_balance.csv")
p <- basinSnowPlot(waterBalance)}
}
\seealso{
\code{\link{read_MESH_OutputTimeseries_csv}} \code{\link{basinStoragePlot}}
}
\author{
Kevin Shook
}
