% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinSnowRainPondedPlot.R
\name{basinSnowRainPondedPlot}
\alias{basinSnowRainPondedPlot}
\title{Plots basin snow and rain in the canopy and ponded water}
\usage{
basinSnowRainPondedPlot(basinWaterBalance)
}
\arguments{
\item{basinWaterBalance}{Required. Data frame to be plotted. As read in by \code{readOutputTimeseriesCSV}.}
}
\value{
Returns a \pkg{ggplot2} line plot of the basin variables (mm).
}
\description{
Plots basin snow and rain in the canopy and ponded water
}
\examples{
\dontrun{
waterBalance <- readOutputTimeseriesCSV("Basin_average_water_balance.csv")
p <- basinSnowRainPondedPlot(waterBalance)}
}
\seealso{
\code{\link{readOutputTimeseriesCSV}} \code{\link{basinStoragePlot}}
}
\author{
Kevin Shook
}
