% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinSoilWaterIcePlot.R
\name{basinSoilWaterIcePlot}
\alias{basinSoilWaterIcePlot}
\title{Plots basin soil water and ice}
\usage{
basinSoilWaterIcePlot(basinWaterBalance, layers = c(1, 2))
}
\arguments{
\item{basinWaterBalance}{Required. Data frame to be plotted. As read in by \code{read_MESH_OutputTimeseries_csv}.}

\item{layers}{Optional. A vector of the layers to be plotted. By default layers 1 and 2 are used.}
}
\value{
Returns a \pkg{ggplot2} line plot of the variable values (mm) faceted
by Layer number. The name of the variable is \code{layer}, so you can change
change the facetting.
}
\description{
Plots basin soil water and ice
}
\examples{
\dontrun{
waterBalance <- read_MESH_OutputTimeseries_csv("Basin_average_water_balance.csv")
p <- basinSoilWaterIcePlot(waterBalance)}
}
\seealso{
\code{\link{read_MESH_OutputTimeseries_csv}} \code{\link{basinStoragePlot}}
}
\author{
Kevin Shook
}
