% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinWaterBalancePlot.R
\name{basinWaterBalancePlot}
\alias{basinWaterBalancePlot}
\title{Plots basin precipitation cumulative water balance.}
\usage{
basinWaterBalancePlot(basinWaterBalance)
}
\arguments{
\item{basinWaterBalance}{Required. Data frame to be plotted. As read in by 
\code{readOutputTimeseriesCSV}. Note that because the value of \code{DTSG} 
(delta storage) can be negative, you need to set a threshold value much smaller than
zero when you read in the values.}
}
\value{
Returns a \pkg{ggplot2} line plot of the variable values (mm).
}
\description{
As with the \code{basinRunoffPlot}, the cumulative values of
precipitation, evaporation and runoff are cumputed by the function, rather
than by using the MESH variables.
}
\examples{
\dontrun{
waterBalance <- readOutputTimeseriesCSV("Basin_average_water_balance.csv", 
missingValueThreshold = -1e6)
p <- basinWaterBalancePlot(waterBalance)}
}
\seealso{
\code{\link{readOutputTimeseriesCSV}} \code{\link{basinStoragePlot}} \code{\link{basinSoilWaterIcePlot}}
}
\author{
Kevin Shook
}
