% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findRecord}
\alias{findRecord}
\title{Finds record in r2c data specified by string}
\usage{
findRecord(recordLines, string)
}
\arguments{
\item{recordLines}{Requred. Vector of lines from r2c file.}

\item{string}{Required. Record name to searh for.}
}
\value{
Returns trimmed record.
}
\description{
Finds record in r2c data specified by string
}
\author{
Kevin Shook
}
