% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridPrecip.R
\name{gridPrecip}
\alias{gridPrecip}
\title{Grids station precipitation}
\usage{
gridPrecip(precip = NULL, shed_raster = NULL, IDW_file = NULL,
  quiet = TRUE)
}
\arguments{
\item{precip}{Required. A list containing 3 elements: 1. the header meta data, 
2. the column meta data, and 3. the precipitation values (in mm). These values
are returned automatically by the \pkg{MESHr} command \code{read_tb0}.}

\item{shed_raster}{Required. A \code{RasterBrick} object describing the MESH basin. 
This can be created using the \pkg{MESHr} command \code{read_r2c_shed} with the 
parameter \code{as_rasters = TRUE}.}

\item{IDW_file}{Required. Output file which holds gridded precipitation for all
time steps.}

\item{quiet}{Optional. If \code{TRUE} (the default) messages are suppressed. 
If \code{FALSE}, the time interval and messages from each gridding are listed.}
}
\value{
If unsuccessful, returns \code{FALSE}. If successful, returns 
\code{TRUE} and the gridded precipitation values are written to the 
\code{IDW_file}. Note that each interval's precipitation is written as it
is gridded. This saves on memory, but can be quite slow.
}
\description{
Grids station interval precipitation values, so that they can
be used as MESH inputs. The gridding method is performed by the \pkg{hydroTSM}
function \code{hydrokrige}, using the IDW (inverse distance weighting) algorithm.
}
\examples{
\dontrun{
hourly_precip_file <- "Red_Deer_all_hourly_precip_new.tb0"
precip <- read_tb0(hourly_precip_file, values_only = FALSE, timezone = "Etc/GMT+7", NAvalue = -0.1)
shedfile <- "RedDeer_MESH_drainage_database.r2c"
shed_raster <- read_r2c_shed(shedfile, as_rasters = TRUE, values_only = TRUE)
IDW_file <- "RedDeerPrecip.idw"
gridPrecip(precip, shed_raster, IDW_file)}
}
