% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MESHr-package.R
\docType{package}
\name{MESHr-package}
\alias{MESHr-package}
\title{Functions for MESH pre- and post- processing}
\description{
The intent of this package is to contain functions do do common tasks
for MESH modelling, such as reading output, plotting, and assessing model quality. The 
first functions will use .csv files - support for netCDF will be added.
}
\references{
To cite \pkg{MESHr} in publications, use the command \code{citation("MESHr")} to get the current version of the citation.\cr
}
