% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_AEP_csv.R
\name{read_AEP_csv}
\alias{read_AEP_csv}
\title{Reads csv file produced by Alberta Environment and Parks}
\usage{
read_AEP_csv(AEPfile = "", timezone = "", values_only = TRUE)
}
\arguments{
\item{AEPfile}{Required. Name of AEP file to be read in.}

\item{timezone}{Optional. The name of the timezone of the data as a character string.
If the \code{timezone} is not specified, your default value (i.e. your
time zone) will be used.  This should be the timezone of your data, but 
omitting daylight savings time. Note that the timezone code is specific 
to your OS. To avoid problems, you should use a timezone without daylight 
savings time. Under Windows or OSX, you can use \option{etc/GMT+6} or 
\option{etc/GMT+7} for Central Standard and Mountain Standard time. Under Linux
you should use \option{Etc/GMT+6} or \option{Etc/GMT+7}.}

\item{values_only}{optional. If \code{TRUE} (the default), only the 
time series values will be returned. If {FALSE}, the meta data will also be 
returned.}
}
\value{
Returns the time series data as a data frame, with the POSIXct 
variable \code{datetime} as the time stamp. Note that the time series
interval may be irregular. If \code{values_only = TRUE}, then
the returned value will be a list conisting of the time series
data frame and the header meta data as a list, with the variables 
\code{values} and \code{header_meta}, respectively. The meta data are:
 \describe{
 \item{variable}{type}
 \item{station_site}{character}
 \item{station_name}{character}
 \item{station_number}{character}
 \item{parameter_name}{character}
 \item{parameter_type}{character}
 \item{parameter_type_name}{character}
 \item{time_series_name}{character}
 \item{time_series_unit}{character}
 \item{longitude}{numeric}
 \item{latitude}{numeric}
 }
}
\description{
Reads csv file produced by Alberta Environment and Parks
}
\examples{
\dontrun{
precip <-  read_AEP_csv("05CA805 Skoki Lodge - PC - C.Merged - All.csv", 
values_only = FALSE)
# show values
head(precip$values)
# show latitude
precip$header_meta$latitude}
}
\seealso{
\code{\link{read_MESH_OutputTimeseries_csv}}
}
\author{
Kevin Shook
}
